Call mix firmware from firmware.burn task.
Also moving original firmware.burn task into a new burn task, which only
burns firmware to an SDCard in case folks want to continue working that
way. This is being done in an effort to avoid confusion for new nerves
users as we've noticed that some users misinterpret this task and think
that it does both.  Hopefully this change will make the tasks more
intuitive based on their task names.
